programación dialog y retorno paquete desde add pruebas

+++ app/src/main/java/com/example/politravel_mariomolina/MapsFragment.kt
@@ -34,7 +34,9 @@ class MapsFragment() : Fragment() {
         {
             googleMap.addMarker(MarkerOptions().position(LatLng(itinerari.latitud,itinerari.longitud)).title(itinerari.nom))
         }
-        googleMap.moveCamera(CameraUpdateFactory.newLatLngZoom(inici,paquet.grausGoogleMaps))
+        googleMap.moveCamera(CameraUpdateFactory.newLatLngZoom(inici,
+            paquet.grausGoogleMaps.toFloat()
+        ))
     }
 
     override fun onCreateView(
